pyproject.toml
[project]
name = "ai-agent-team"
version = "0.1.0"
description = "Multi-agent team system using Anthropic SDK"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "python-dotenv>=1.0.0",
    "PyQt6>=6.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["E", "F", "I"]
